requirements.txt
pycurl
MySQL-python >= 1.2.3
ujson
python-memcached
